+++ docker/docker-compose.yml
@@ -1,0 +1,24 @@
+version: '3.8'
+
+services:
+  web:
+    build: . 
+    container_name: destiny_calculator_app
+    ports:
+      - "8080:8080" 
+    volumes:
+      - ./instance:/app/instance # Сохранение данных БД при перезапуске
+    environment:
+      PYTHONUNBUFFERED: 1
+
+  locust:
+    build: . 
+    container_name: destiny_calculator_locust
+    profiles: ["test"]
+    ports:
+      - "8089:8089"
+    volumes:
+      - .:/app
+    command: ["locust", "-f", "load_test.py", "--host=http://web:8080"]
+    depends_on:
+      - web
